% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGaussianFiles.R
\name{readgausscalctime}
\alias{readgausscalctime}
\title{Read calculation time from Gaussian 16 output}
\usage{
readgausscalctime(Path, logcont = NULL)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}

\item{logcont}{The content of a .log file as read in by readLines(). May be
used to combine multiple readgauss functions by avoiding to re-read the
file in each function.}
}
\value{
A numeric vector containing the calculation time in seconds (s).
}
\description{
This function reads the calculation time from a Gaussian 16 (and potentially
earlier) output file.
}
\details{
readgausscalctime is designed to read the calculation time from a Gaussian 16
output file. It can potentially work with earlier versions of Gaussian as
well. The function returns the Job CPU Time in s.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readgausscalctime("Path/To/File.out")
}
  
}
