% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGaussianFiles.R
\name{readgaussdipole}
\alias{readgaussdipole}
\title{Read dipole moment from Gaussian 16 output}
\usage{
readgaussdipole(Path, logcont = NULL)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}

\item{logcont}{The content of a .log file as read in by readLines(). May be
used to combine multiple readgauss functions by avoiding to re-read the
file in each function.}
}
\value{
A numeric vector containing the dipole moments in Debye (D).
}
\description{
This function reads the total dipole moment as well as the X, Y and Z
components of the dipole moment from a Gaussian 16 (and potentially earlier)
output file.
}
\details{
\code{readgaussdipole} is designed to read the total dipole moment as well as the
X, Y and Z components of the dipole moment from a Gaussian 16
output file. It can potentially work with earlier versions of Gaussian as
well. The function returns the dipole moments in Debye.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readgaussdipole("Path/To/File.out")
}

}
