% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGaussianFiles.R
\name{readgaussener}
\alias{readgaussener}
\title{Read energies from Gaussian 16 output}
\usage{
readgaussener(Path, logcont = NULL)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}

\item{logcont}{The content of a .log file as read in by readLines(). May be
used to combine multiple readgauss functions by avoiding to re-read the
file in each function.}
}
\value{
A numeric vector containing the energy values in kilojoules per mole
  (kJ/mol).
}
\description{
This function reads energy data from a Gaussian 16 (and potentially earlier)
output file.
}
\details{
\code{readgaussener} is designed to read energy data from a Gaussian 16 output file.
It can potentially work with earlier versions of Gaussian as well. The
function returns the Zero-point Energy, the Energy, the Enthalpies as well as
the Free Energies in kJ/mol.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readgaussener("Path/To/File.out")
}

}
