% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGaussianFiles.R
\name{readgaussfroorb}
\alias{readgaussfroorb}
\title{Read frontier orbital energies from Gaussian 16 output}
\usage{
readgaussfroorb(Path, logcont = NULL)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}

\item{logcont}{The content of a .log file as read in by readLines(). May be
used to combine multiple readgauss functions by avoiding to re-read the
file in each function.}
}
\value{
A numeric vector containing the energy values in kilojoules per mole
  (kJ/mol).
}
\description{
This function reads the frontier orbital energies of the Highest Occupied
Molecular Orbital (HOMO) and the Lowest Unoccupied Molecular Orbital (LUMO)
from a Gaussian 16 (and potentially earlier) output file.
}
\details{
\code{readgaussfroorb} is designed to read the frontier orbital energies of the
Highest Occupied Molecular Orbital (HOMO) and the Lowest Unoccupied Molecular
Orbital (LUMO) from a Gaussian 16. It can potentially work with earlier
versions of Gaussian as well. It also calculates the HOMO-LUMO energy
difference. The function returns the dipole moments in in kJ/mol and assumes
at max 5 orbitals per line in the output file.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readgaussfroorb("Path/To/File.out")
}

}
