% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGaussianFiles.R
\name{readgaussgeom}
\alias{readgaussgeom}
\title{Read geometry and Mullikan charges from Gaussian 16 output}
\usage{
readgaussgeom(Path, logcont = NULL)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}

\item{logcont}{The content of a .log file as read in by readLines(). May be
used to combine multiple readgauss functions by avoiding to re-read the
file in each function.}
}
\value{
A data frame containing the extracted information from the Gaussian
  output file. The data frame includes the following columns:
  \describe{
  \item{Num}{The atom number.}
  \item{AtomNum}{The atomic number of the atom.}
  \item{Sym}{The atomic symbol of the atom.}
  \item{Lab}{Unique atom label assigned to each atom.}
  \item{Type}{The atom type.}
  \item{X}{The X-coordinate of the atom in the Cartesian coordinate system.}
  \item{Y}{The Y-coordinate of the atom in the Cartesian coordinate system.}
  \item{Z}{The Z-coordinate of the atom in the Cartesian coordinate system.}
  \item{MullikanCharge}{The Mullikan charge associated with each atom.}
  }
}
\description{
This function reads geometric and Mullikan charges data from a Gaussian 16
(and potentially earlier) output file.
}
\details{
\code{readgaussgeom} is designed to extract geometric and Mullikan charges data from
a Gaussian 16 output file, and potentially earlier versions. Gaussian 16 is a
widely used quantum chemistry software package for electronic structure
calculations.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readgaussgeom("Path/To/File.out")
}

}
