% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGaussianFiles.R
\name{readgaussout}
\alias{readgaussout}
\title{Read vibrational data from Gaussian 09 output}
\usage{
readgaussout(Path, logcont = NULL)
}
\arguments{
\item{Path}{Name (with path) of the file from which the data are to be read.}

\item{logcont}{The content of a .log file as read in by readLines(). This
parameter can be used to combine multiple readgaussout function calls by
avoiding re-reading the file in each function. If logcont is provided, the
function uses it as the content of the file instead of reading from Path.}
}
\value{
A matrix with three rows containing the vibrational frequency/band
  position in cm-1 (\code{BPosi}), the IR intensity (\code{IRint}) as well as
  the Raman activity (\code{RActi}) for each vibration.
}
\description{
This function reads vibrational data from a Gaussian 09 (and potentially
later) output file.
}
\details{
\code{readgaussout} reads vibrational data from a Gaussian 09 (and potentially
later) output file. It extracts the frequencies, IR intensities, and Raman
activities from the file and returns them as a matrix.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readgaussout("Path/To/File.out")
}

}
