% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadGaussianFiles.R
\name{readgaussvibmotion}
\alias{readgaussvibmotion}
\title{Read atom motion during vibration from Gaussian 16 output}
\usage{
readgaussvibmotion(Path, logcont = NULL)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}

\item{logcont}{The content of a .log file as read in by readLines(). May be
used to combine multiple readgauss functions by avoiding to re-read the
file in each function.}
}
\value{
A data frame containing the atom motion during vibrations.
  \describe{
    \item{Num}{The atom number.}
    \item{AtomNum}{The atomic number of the atom.}
    \item{Sym}{The atomic symbol of the atom.}
    \item{Lab}{Unique atom label assigned to each atom.}
    \item{Vib}{Vibration index.}
    \item{dX}{Displacement in the X direction.}
    \item{dY}{Displacement in the Y direction.}
    \item{dZ}{Displacement in the Z direction.}
  }
}
\description{
This function reads the atom motions from a Gaussian 16 (and potentially
earlier) output file.
}
\details{
\code{readgaussvibmotion} is designed to read the calculated motions of individual
atoms during vibrations from a Gaussian 16 output file. The function can
potentially work with earlier versions of Gaussian as well. The function
assumes 3N-6 vibrations, where N is the number of atoms. It might give
strange results for linear molecules with 3N-5 vibrations.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readgaussvibmotion("Path/To/File.out")
}

}
