% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadOrcaFiles.R
\name{readorcacalctime}
\alias{readorcacalctime}
\title{Read calculation time from ORCA 5.0 output}
\usage{
readorcacalctime(Path)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}
}
\value{
A numeric vector containing the calculation time in seconds (s).
}
\description{
This function reads the calculation time from a ORCA 5.0 (and potentially
earlier) output file.
}
\details{
\code{readorcacalctime} is designed to read the calculation time from a ORCA 5.0
output file. It can potentially work with earlier and later versions of ORCA
as well. The function returns the Total Run Time in s.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readorcacalctime("Path/To/File.out")
}

}
