% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadOrcaFiles.R
\name{readorcaener}
\alias{readorcaener}
\title{Read energies from ORCA 5.0 output}
\usage{
readorcaener(Path)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}
}
\value{
A numeric vector containing the energy values in kilojoules per mole
  (kJ/mol).
}
\description{
This function reads energy data from a ORCA 5.0 (and potentially later)
output file.
}
\details{
\code{readorcaener} is designed to read energy data from a ORCA 5.0 output file.
It can potentially work with earlier and later versions of ORCA as well. The
function returns the Electronic Energy, the Total Thermal Energy, the Total
Enthalpy as well as the Final Gibbs Free Energies in kJ/mol.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readorcaener("Path/To/File.out")
}

}
