% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadOrcaFiles.R
\name{readorcaout}
\alias{readorcaout}
\title{Read vibrational data from ORCA 5.0 output}
\usage{
readorcaout(Path)
}
\arguments{
\item{Path}{Name (with path) of the file which the data are to be read from.}
}
\value{
A matrix with three rows containing the vibrational frequency/band
  position in cm-1 (\code{BPosi}), the IR intensity (\code{IRint}) as well as
  the Raman activity (\code{RActi}) for each vibration.
}
\description{
This function reads vibrational data from a ORCA 5.0 (and potentially
later) output file.
}
\details{
\code{readorcaout} reads vibrational data from a ORCA 5.0 (and potentially
later) output file. It extracts the frequencies, IR intensities, and Raman
activities from the file and returns them as a matrix.
}
\examples{
\dontrun{
  ## NOT RUN
  QMout <- readorcaout("Path/To/File.out")
}

}
