% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readxyz.R
\name{readxyz}
\alias{readxyz}
\title{Read xyz file}
\usage{
readxyz(path, skip = 2, ...)
}
\arguments{
\item{path}{The name (with path) of the file to be read.}

\item{skip}{The number of lines to skip during the read process. Default is
2, assuming the file contains a header and the number of atoms.}

\item{...}{Additional arguments passed to the underlying read.table function.}
}
\value{
\code{readxyz} constructs a \code{data.frame} with seven columns for
  the number of each atom (\code{Num}), atomic numbers (\code{AtomNum}), atom
  symbols (\code{Sym}), unique labels (\code{Lab}), and the Cartesian
  coordinates (\code{X}, \code{Y}, \code{Z}) from the read data frame. The
  column names are set accordingly.
}
\description{
This function reads data from an .xyz file and returns a data frame with the
parsed information.
}
\details{
This function reads data from an .xyz file and returns a data frame with the
parsed information. .xyz files are commonly used to represent molecular
structures, where each line contains the atomic symbol followed by its
Cartesian coordinates.
}
\examples{
\dontrun{
  ## NOT RUN
  XYZout <- readxyz("Path/To/File.xyz")
}

}
