% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueAtomLabels.R
\name{uniqueAtomLabels}
\alias{uniqueAtomLabels}
\title{Create unique alphanumeric atom labels}
\usage{
uniqueAtomLabels(label)
}
\arguments{
\item{label}{Vector containing the atom labels}
}
\description{
This function takes a vector of atom labels and returns a vector with a
unique alphanumeric label for each atom.
}
\details{
This function takes a vector of atom labels and returns a vector with unique
alphanumeric labels. If the input vector contains numeric labels, the
function converts them to their corresponding symbol using the
\code{\link[PeriodicTable]{symb}} function.
}
\examples{
  examlab <- c("C", "H", "H", "H", "H")
  unilab <- uniqueAtomLabels(examlab)

}
